sensor/requirements.txt
paho-mqtt==0.4.90